fix: del some transformation
