+++ download_nasa_apod_photo.py
@@ -1,24 +1,20 @@
 import logging
 import os
 import requests
+import time
+
 from pathlib import Path
-from urllib.parse import urlsplit
 from dotenv import load_dotenv
+
+from python_utils import get_file_extension
+
 
 API = "https://api.nasa.gov/planetary/apod"
 
 
-def get_file_extension(url: str) -> str:
-    """Function get http link and return file extension"""
-    url = urlsplit(url)[2]
-    url_link = os.path.splitext(url)
-    file_link, *extension = url_link
-    return extension[0]
-
-
 def get_nasa_photo(api: str, token: str, count: int) -> list:
     """Take nasa token, API, count -> return links"""
-    params = {"api_key": f"{token}", "count": f"{count}"}
+    params = {"api_key": token, "count": count}
     response = requests.get(api, params=params)
     response.raise_for_status()
     photo_link = response.json()
@@ -41,13 +37,16 @@ def download_images(url: str, number: int, extension: str) -> object:
 
 
 if __name__ == "__main__":
+    start_time = time.time()
     load_dotenv()
     token = os.environ["NASA_API_TOKEN"]
     download_path = os.getenv("DOWNLOAD_PATH")
     Path(download_path).mkdir(parents=True, exist_ok=True)
-    nasa_link = get_nasa_photo(API, token, 50)
+    nasa_link = get_nasa_photo(API, token, 45)
 
     for number, link in enumerate(nasa_link):
         extension = get_file_extension(link)
         download_images(link, number, extension)
     logging.debug("Function: download_images - Done")
+    end_time = time.time() - start_time
+    print(end_time)
